Removing remnants of the cmake buildsystem.
CMake is a largely undocumented mess, that's hard to pick up for new people
or anyone that hasn't looked at it in a few years.  The same seems true for
for qmake and QBS.  So I've opted to use QBS because out of all of the bad
options this one seems to hold the most promise.  It has a nicer syntax.
That's pretty much it.